New training with more epochs
The new training shows good performance with improvements in some metrics.

The accuracy has progressively improved and has reached a value close to 1.0 in the last epochs. This indicates that the model is correctly classifying most of the images in the validation set.

The error_rate has reduced and remains low, which is also a good sign.

The precision_score is high and remains constant across all epochs, indicating that the model is minimizing false positives and correctly classifying positive images.

The recall_score is high and remains constant across all epochs, meaning that the model is correctly capturing the majority of positive images.

The f1_score is high and remains constant across all epochs, indicating a balance between model accuracy and recall.

Overall, the new training shows promising performance, with high accuracy, recall and F1 metrics. However, further evaluation will be performed and will validate the model with independent data to get a complete picture of its performance.